+++ R/scale_hdx.R
@@ -15,7 +15,7 @@ scale_color_hdx <- function(...) {
   ggplot2::discrete_scale(
     "colour",
     "hdx",
-    hdx_pal(),
+    hdx_pal_discrete(),
     na.value = hdx_hex("gray-light"),
     ...
   )
@@ -24,6 +24,30 @@ scale_color_hdx <- function(...) {
 #' @rdname scale_hdx
 #' @export
 scale_colour_hdx <- scale_color_hdx
+
+#' @rdname scale_hdx
+#' @export
+scale_color_hdx_gray <- function(...) {
+  ggplot2::discrete_scale(
+    "colour",
+    "hdx",
+    hdx_pal_gray(),
+    na.value = hdx_hex("tomato-hdx"),
+    ...
+  )
+}
+
+#' @rdname scale_hdx
+#' @export
+scale_colour_hdx_gray <- scale_color_hdx_gray
+
+#' @rdname scale_hdx
+#' @export
+scale_colour_hdx_grey <- scale_color_hdx_gray
+
+#' @rdname scale_hdx
+#' @export
+scale_color_hdx_grey <- scale_color_hdx_gray
 
 #' @rdname scale_hdx
 #' @export
@@ -79,11 +103,27 @@ scale_fill_hdx <- function(...) {
   ggplot2::discrete_scale(
     "fill",
     "hdx",
-    hdx_pal(),
+    hdx_pal_discrete(),
     na.value = hdx_hex("gray-light"),
     ...
   )
 }
+
+#' @rdname scale_hdx
+#' @export
+scale_fill_hdx_gray <- function(...) {
+  ggplot2::discrete_scale(
+    "fill",
+    "hdx",
+    hdx_pal_gray(),
+    na.value = hdx_hex("tomato-hdx"),
+    ...
+  )
+}
+
+#' @rdname scale_hdx
+#' @export
+scale_fill_hdx_grey <- scale_fill_hdx_gray
 
 #' @rdname scale_hdx
 #' @export
@@ -128,8 +168,8 @@ scale_fill_gradient_hdx <- function(...) {
     aesthetics = "fill",
     scale_name = "hdx",
     palette = scales::seq_gradient_pal(
-      low = hdx_colors()["tomato-hdx"],
-      high = hdx_colors()["sapphire-hdx"]
+      low = hdx_hex("tomato-hdx"),
+      high = hdx_hex("sapphire-hdx")
     ),
     guide = "colorbar",
     ...
@@ -143,8 +183,8 @@ scale_fill_gradient_hdx_sapphire <- function(...) {
     aesthetics = "fill",
     scale_name = "hdx",
     palette = scales::seq_gradient_pal(
-      low = hdx_colors()["sapphire-ultra-light"],
-      high = hdx_colors()["sapphire-hdx"]
+      low = "white",
+      high = hdx_hex("sapphire-hdx")
     ),
     guide = "colorbar",
     ...
@@ -158,8 +198,8 @@ scale_fill_gradient_hdx_mint <- function(...) {
     aesthetics = "fill",
     scale_name = "hdx",
     palette = scales::seq_gradient_pal(
-      low = hdx_colors()["mint-ultra-light"],
-      high = hdx_colors()["mint-hdx"]
+      low = "white",
+      high = hdx_hex("mint-hdx")
     ),
     guide = "colorbar",
     ...
@@ -173,8 +213,8 @@ scale_fill_gradient_hdx_tomato <- function(...) {
     aesthetics = "fill",
     scale_name = "hdx",
     palette = scales::seq_gradient_pal(
-      low = hdx_colors()["tomato-ultra-light"],
-      high = hdx_colors()["tomato-hdx"]
+      low = "white",
+      high = hdx_hex("tomato-hdx")
     ),
     guide = "colorbar",
     ...
@@ -188,8 +228,8 @@ scale_color_gradient_hdx <- function(...) {
     aesthetics = "color",
     scale_name = "hdx",
     palette = scales::seq_gradient_pal(
-      low = hdx_colors()["tomato-hdx"],
-      high = hdx_colors()["sapphire-hdx"]
+      low = hdx_hex("tomato-hdx"),
+      high = hdx_hex("sapphire-hdx")
     ),
     guide = "colorbar",
     ...
@@ -207,8 +247,8 @@ scale_color_gradient_hdx_sapphire <- function(...) {
     aesthetics = "color",
     scale_name = "hdx",
     palette = scales::seq_gradient_pal(
-      low = hdx_colors()["sapphire-ultra-light"],
-      high = hdx_colors()["sapphire-hdx"]
+      low = "white",
+      high = hdx_hex("sapphire-hdx")
     ),
     guide = "colorbar",
     ...
@@ -226,8 +266,8 @@ scale_color_gradient_hdx_mint <- function(...) {
     aesthetics = "color",
     scale_name = "hdx",
     palette = scales::seq_gradient_pal(
-      low = hdx_colors()["mint-ultra-light"],
-      high = hdx_colors()["mint-hdx"]
+      low = "white",
+      high = hdx_hex("mint-hdx")
     ),
     guide = "colorbar",
     ...
@@ -245,8 +285,8 @@ scale_color_gradient_hdx_tomato <- function(...) {
     aesthetics = "color",
     scale_name = "hdx",
     palette = scales::seq_gradient_pal(
-      low = hdx_colors()["tomato-ultra-light"],
-      high = hdx_colors()["tomato-hdx"]
+      low = "white",
+      high = hdx_hex("tomato-hdx")
     ),
     guide = "colorbar",
     ...
@@ -264,9 +304,9 @@ scale_color_gradient2_hdx <- function(...) {
     aesthetics = "color",
     scale_name = "hdx",
     palette = scales::div_gradient_pal(
-      low = hdx_colors()["tomato-hdx"],
-      mid = hdx_colors()["gray-white"],
-      high = hdx_colors()["sapphire-hdx"]
+      low = hdx_hex("tomato-hdx"),
+      mid = hdx_hex("gray-white"),
+      high = hdx_hex("sapphire-hdx")
     ),
     guide = "colorbar",
     ...
@@ -284,9 +324,9 @@ scale_fill_gradient2_hdx <- function(...) {
     aesthetics = "fill",
     scale_name = "hdx",
     palette = scales::div_gradient_pal(
-      low = hdx_colors()["tomato-hdx"],
-      mid = hdx_colors()["gray-white"],
-      high = hdx_colors()["sapphire-hdx"]
+      low = hdx_hex("tomato-hdx"),
+      mid = hdx_hex("gray-white"),
+      high = hdx_hex("sapphire-hdx")
     ),
     guide = "colorbar",
     ...