+++ R/utils-na.R
@@ -1,3 +1,0 @@
-#' Return NA if single NA passed in
-#'
-#' @noRd