+++ tests/testthat/test-hdx_pal.R
@@ -1,12 +1,12 @@
-test_that("hdx_pal() returns hex codes", {
-  pal <- hdx_pal()
+test_that("hdx_pal_discrete() returns hex codes", {
+  pal <- hdx_pal_discrete()
   for (i in 1:12) {
     expect_hex_colors(pal(i))
   }
 })
 
-test_that("hdx_pal() raises correctly", {
-  pal <- hdx_pal()
+test_that("hdx_pal_discrete() raises correctly", {
+  pal <- hdx_pal_discrete()
   expect_error(pal(0))
   expect_warning(pal(13))
 })
@@ -15,7 +15,8 @@ test_that("hdx_pal_color()s return hex codes", {
   pals <- list(
     hdx_pal_mint(),
     hdx_pal_sapphire(),
-    hdx_pal_tomato()
+    hdx_pal_tomato(),
+    hdx_pal_gray()
   )
 
   for (pal in pals) {
@@ -25,7 +26,7 @@ test_that("hdx_pal_color()s return hex codes", {
   }
 })
 
-test_that("hdx_pal() raises errors", {
+test_that("hdx_pal_...() raises errors", {
   pals <- list(
     hdx_pal_mint(),
     hdx_pal_sapphire(),
@@ -38,7 +39,7 @@ test_that("hdx_pal() raises errors", {
   }
 })
 
-test_that("hdx_pal() raises warnings", {
-  pal <- hdx_pal()
+test_that("hdx_pal_discrete() raises warnings", {
+  pal <- hdx_pal_discrete()
   expect_warning(pal(13))
 })