+++ NEWS.md
@@ -1,8 +1,17 @@
 # Development
 
-## gghdx 0.0.2.2
+## gghdx 0.1
 
-* Right adjusting y-axis labels in `theme_hdx()`. 
+* Right adjusting y-axis labels in `theme_hdx()`.
+* Replace deprecated `size` argument from `ggplot2` calls.
+* Move `ultra-light` colors to end of `hdx_pal()` discrete palette.
+* Make output of `hdx_hex()` unnamed to avoid errors in `ggplot2` and
+allow both grey and gray spellings of colors.
+* Change `..._gradient_...` scale functions to have `white` as the low color.
+* Add `display_hdx_pal()` function to allow visualizing the HDX color palettes.
+* Improved error catching when fonts are not available.
+* Added gray color palette, `hdx_pal_gray()`.
+
 
 ## gghdx 0.0.2.1